frontend/requirements.txt
streamlit
requests
pandas
plotly
python-dotenv

